pyproject.toml
[tool.poetry]
name = "dietapp"
version = "0.1.0"
description = ""
authors = ["Arkadiusz Tarnowski"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
django-environ = "^0.10.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
